pyproject.toml
[project]
name = "my-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.5.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
]
